refactoring+full_voicing
