+++ game/Submods/AI_submod/monikai_emotion.rpy
@@ -1,0 +1,18 @@
+#Face emotion recognition (not available yet)
+
+$ emotion_list = ["anger","disgust","fear","joy","sadness","surprise","neutral"]
+
+define negative_emotions = ["anger", "disgust", "fear", "sadness"]
+define positive_emotions = ["joy", "surprise"]
+define neutral_emotions = ["neutral"]
+
+define sentences_emotions = {
+    "angry": "I'm sorry my love, I didn't mean to upset you.",
+    "disgusted": "What is going on ? If you feel like there is something wrong, please tell me.",
+    "fearful": "Don't be afraid sweetie, I'm here for you.",
+    "happy": "I'm glad you're happy darling.",
+    "neutral": "Let's do something fun together!",
+    "sad": "If you feel down, always remember that I'm here for you. I love you.",
+    "surprised": "What is it ? Is there something wrong ?",
+    "no": "Oh it seems that I can't see through your camera sweetie. Maybe face me a little bit more ?",
+}
