+++ old_scripts/speech_to_text.py
