+++ setup.sh
@@ -1,8 +1,4 @@
 #!/bin/bash
 
 pip install -r requirements.txt
-git clone https://github.com/coqui-ai/TTS
-cd TTS
-pip install -e .
-cd ..
-python -m playwright install
+python -m playwright install firefox
